merge upstream

+++ src/Libraries/Nop.Services/Directory/CurrencyService.cs
@@ -1,11 +1,10 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using Nop.Core;
 using Nop.Core.Caching;
 using Nop.Core.Data;
 using Nop.Core.Domain.Directory;
-using Nop.Core.Plugins;
 using Nop.Services.Events;
 using Nop.Services.Stores;
 
@@ -16,81 +15,39 @@ namespace Nop.Services.Directory
     /// </summary>
     public partial class CurrencyService : ICurrencyService
     {
-        #region Constants
-        /// <summary>
-        /// Key for caching
-        /// </summary>
-        /// <remarks>
-        /// {0} : currency ID
-        /// </remarks>
-        private const string CURRENCIES_BY_ID_KEY = "Nop.currency.id-{0}";
-        /// <summary>
-        /// Key for caching
-        /// </summary>
-        /// <remarks>
-        /// {0} : show hidden records?
-        /// </remarks>
-        private const string CURRENCIES_ALL_KEY = "Nop.currency.all-{0}";
-        /// <summary>
-        /// Key pattern to clear cache
-        /// </summary>
-        private const string CURRENCIES_PATTERN_KEY = "Nop.currency.";
-
-        #endregion
-
         #region Fields
 
-        private readonly IRepository<Currency> _currencyRepository;
-        private readonly IStoreMappingService _storeMappingService;
-        private readonly ICacheManager _cacheManager;
         private readonly CurrencySettings _currencySettings;
-        private readonly IPluginFinder _pluginFinder;
         private readonly IEventPublisher _eventPublisher;
+        private readonly IExchangeRatePluginManager _exchangeRatePluginManager;
+        private readonly IRepository<Currency> _currencyRepository;
+        private readonly IStaticCacheManager _cacheManager;
+        private readonly IStoreMappingService _storeMappingService;
 
         #endregion
 
         #region Ctor
 
-        /// <summary>
-        /// Ctor
-        /// </summary>
-        /// <param name="cacheManager">Cache manager</param>
-        /// <param name="currencyRepository">Currency repository</param>
-        /// <param name="storeMappingService">Store mapping service</param>
-        /// <param name="currencySettings">Currency settings</param>
-        /// <param name="pluginFinder">Plugin finder</param>
-        /// <param name="eventPublisher">Event published</param>
-        public CurrencyService(ICacheManager cacheManager,
+        public CurrencyService(CurrencySettings currencySettings,
+            IEventPublisher eventPublisher,
+            IExchangeRatePluginManager exchangeRatePluginManager,
             IRepository<Currency> currencyRepository,
-            IStoreMappingService storeMappingService,
-            CurrencySettings currencySettings,
-            IPluginFinder pluginFinder,
-            IEventPublisher eventPublisher)
+            IStaticCacheManager cacheManager,
+            IStoreMappingService storeMappingService)
         {
-            this._cacheManager = cacheManager;
-            this._currencyRepository = currencyRepository;
-            this._storeMappingService = storeMappingService;
-            this._currencySettings = currencySettings;
-            this._pluginFinder = pluginFinder;
-            this._eventPublisher = eventPublisher;
+            _currencySettings = currencySettings;
+            _eventPublisher = eventPublisher;
+            _exchangeRatePluginManager = exchangeRatePluginManager;
+            _currencyRepository = currencyRepository;
+            _cacheManager = cacheManager;
+            _storeMappingService = storeMappingService;
         }
 
         #endregion
-        
+
         #region Methods
 
-        /// <summary>
-        /// Gets currency live rates
-        /// </summary>
-        /// <param name="exchangeRateCurrencyCode">Exchange rate currency code</param>
-        /// <returns>Exchange rates</returns>
-        public virtual IList<ExchangeRate> GetCurrencyLiveRates(string exchangeRateCurrencyCode)
-        {
-            var exchangeRateProvider = LoadActiveExchangeRateProvider();
-            if (exchangeRateProvider == null)
-                throw new Exception("Active exchange rate provider cannot be loaded");
-            return exchangeRateProvider.GetCurrencyLiveRates(exchangeRateCurrencyCode);
-        }
+        #region Currency
 
         /// <summary>
         /// Deletes currency
@@ -99,11 +56,14 @@ namespace Nop.Services.Directory
         public virtual void DeleteCurrency(Currency currency)
         {
             if (currency == null)
-                throw new ArgumentNullException("currency");
-            
+                throw new ArgumentNullException(nameof(currency));
+
+            if (currency is IEntityForCaching)
+                throw new ArgumentException("Cacheable entities are not supported by Entity Framework");
+
             _currencyRepository.Delete(currency);
 
-            _cacheManager.RemoveByPattern(CURRENCIES_PATTERN_KEY);
+            _cacheManager.RemoveByPrefix(NopDirectoryDefaults.CurrenciesPrefixCacheKey);
 
             //event notification
             _eventPublisher.EntityDeleted(currency);
@@ -113,26 +73,44 @@ namespace Nop.Services.Directory
         /// Gets a currency
         /// </summary>
         /// <param name="currencyId">Currency identifier</param>
+        /// <param name="loadCacheableCopy">A value indicating whether to load a copy that could be cached (workaround until Entity Framework supports 2-level caching)</param>
         /// <returns>Currency</returns>
-        public virtual Currency GetCurrencyById(int currencyId)
+        public virtual Currency GetCurrencyById(int currencyId, bool loadCacheableCopy = true)
         {
             if (currencyId == 0)
                 return null;
-            
-            string key = string.Format(CURRENCIES_BY_ID_KEY, currencyId);
-            return _cacheManager.Get(key, () => _currencyRepository.GetById(currencyId));
+
+            Currency loadCurrencyFunc()
+            {
+                return _currencyRepository.GetById(currencyId);
+            }
+
+            if (!loadCacheableCopy)
+                return loadCurrencyFunc();
+
+            //cacheable copy
+            var key = string.Format(NopDirectoryDefaults.CurrenciesByIdCacheKey, currencyId);
+            return _cacheManager.Get(key, () =>
+            {
+                var currency = loadCurrencyFunc();
+                if (currency == null)
+                    return null;
+                return new CurrencyForCaching(currency);
+            });
         }
 
         /// <summary>
         /// Gets a currency by code
         /// </summary>
         /// <param name="currencyCode">Currency code</param>
+        /// <param name="loadCacheableCopy">A value indicating whether to load a copy that could be cached (workaround until Entity Framework supports 2-level caching)</param>
         /// <returns>Currency</returns>
-        public virtual Currency GetCurrencyByCode(string currencyCode)
+        public virtual Currency GetCurrencyByCode(string currencyCode, bool loadCacheableCopy = true)
         {
-            if (String.IsNullOrEmpty(currencyCode))
+            if (string.IsNullOrEmpty(currencyCode))
                 return null;
-            return GetAllCurrencies(true).FirstOrDefault(c => c.CurrencyCode.ToLower() == currencyCode.ToLower());
+            return GetAllCurrencies(true, loadCacheableCopy: loadCacheableCopy)
+                .FirstOrDefault(c => c.CurrencyCode.ToLower() == currencyCode.ToLower());
         }
 
         /// <summary>
@@ -140,18 +118,36 @@ namespace Nop.Services.Directory
         /// </summary>
         /// <param name="showHidden">A value indicating whether to show hidden records</param>
         /// <param name="storeId">Load records allowed only in a specified store; pass 0 to load all records</param>
+        /// <param name="loadCacheableCopy">A value indicating whether to load a copy that could be cached (workaround until Entity Framework supports 2-level caching)</param>
         /// <returns>Currencies</returns>
-        public virtual IList<Currency> GetAllCurrencies(bool showHidden = false, int storeId = 0)
+        public virtual IList<Currency> GetAllCurrencies(bool showHidden = false, int storeId = 0, bool loadCacheableCopy = true)
         {
-            string key = string.Format(CURRENCIES_ALL_KEY, showHidden);
-            var currencies = _cacheManager.Get(key, () =>
+            IList<Currency> loadCurrenciesFunc()
             {
                 var query = _currencyRepository.Table;
                 if (!showHidden)
                     query = query.Where(c => c.Published);
-                query = query.OrderBy(c => c.DisplayOrder);
+                query = query.OrderBy(c => c.DisplayOrder).ThenBy(c => c.Id);
                 return query.ToList();
-            });
+            }
+
+            IList<Currency> currencies;
+            if (loadCacheableCopy)
+            {
+                //cacheable copy
+                var key = string.Format(NopDirectoryDefaults.CurrenciesAllCacheKey, showHidden);
+                currencies = _cacheManager.Get(key, () =>
+                {
+                    var result = new List<Currency>();
+                    foreach (var currency in loadCurrenciesFunc())
+                        result.Add(new CurrencyForCaching(currency));
+                    return result;
+                });
+            }
+            else
+            {
+                currencies = loadCurrenciesFunc();
+            }
 
             //store mapping
             if (storeId > 0)
@@ -160,6 +156,7 @@ namespace Nop.Services.Directory
                     .Where(c => _storeMappingService.Authorize(c, storeId))
                     .ToList();
             }
+
             return currencies;
         }
 
@@ -170,11 +167,14 @@ namespace Nop.Services.Directory
         public virtual void InsertCurrency(Currency currency)
         {
             if (currency == null)
-                throw new ArgumentNullException("currency");
+                throw new ArgumentNullException(nameof(currency));
+
+            if (currency is IEntityForCaching)
+                throw new ArgumentException("Cacheable entities are not supported by Entity Framework");
 
             _currencyRepository.Insert(currency);
 
-            _cacheManager.RemoveByPattern(CURRENCIES_PATTERN_KEY);
+            _cacheManager.RemoveByPrefix(NopDirectoryDefaults.CurrenciesPrefixCacheKey);
 
             //event notification
             _eventPublisher.EntityInserted(currency);
@@ -187,17 +187,39 @@ namespace Nop.Services.Directory
         public virtual void UpdateCurrency(Currency currency)
         {
             if (currency == null)
-                throw new ArgumentNullException("currency");
+                throw new ArgumentNullException(nameof(currency));
+
+            if (currency is IEntityForCaching)
+                throw new ArgumentException("Cacheable entities are not supported by Entity Framework");
 
             _currencyRepository.Update(currency);
 
-            _cacheManager.RemoveByPattern(CURRENCIES_PATTERN_KEY);
+            _cacheManager.RemoveByPrefix(NopDirectoryDefaults.CurrenciesPrefixCacheKey);
 
             //event notification
             _eventPublisher.EntityUpdated(currency);
         }
 
+        #endregion
 
+        #region Conversions
+
+        /// <summary>
+        /// Gets live rates regarding the passed currency
+        /// </summary>
+        /// <param name="currencyCode">Currency code; pass null to use primary exchange rate currency</param>
+        /// <returns>Exchange rates</returns>
+        public virtual IList<ExchangeRate> GetCurrencyLiveRates(string currencyCode = null)
+        {
+            var exchangeRateProvider = _exchangeRatePluginManager.LoadPrimaryPlugin()
+                ?? throw new Exception("Active exchange rate provider cannot be loaded");
+
+            currencyCode = currencyCode
+                ?? GetCurrencyById(_currencySettings.PrimaryExchangeRateCurrencyId)?.CurrencyCode
+                ?? throw new NopException("Primary exchange rate currency is not set");
+
+            return exchangeRateProvider.GetCurrencyLiveRates(currencyCode);
+        }
 
         /// <summary>
         /// Converts currency
@@ -222,19 +244,20 @@ namespace Nop.Services.Directory
         public virtual decimal ConvertCurrency(decimal amount, Currency sourceCurrencyCode, Currency targetCurrencyCode)
         {
             if (sourceCurrencyCode == null)
-                throw new ArgumentNullException("sourceCurrencyCode");
+                throw new ArgumentNullException(nameof(sourceCurrencyCode));
 
             if (targetCurrencyCode == null)
-                throw new ArgumentNullException("targetCurrencyCode");
+                throw new ArgumentNullException(nameof(targetCurrencyCode));
 
-            decimal result = amount;
+            var result = amount;
             if (sourceCurrencyCode.Id == targetCurrencyCode.Id)
                 return result;
-            if (result != decimal.Zero && sourceCurrencyCode.Id != targetCurrencyCode.Id)
-            {
-                result = ConvertToPrimaryExchangeRateCurrency(result, sourceCurrencyCode);
-                result = ConvertFromPrimaryExchangeRateCurrency(result, targetCurrencyCode);
-            }
+
+            if (result == decimal.Zero || sourceCurrencyCode.Id == targetCurrencyCode.Id)
+                return result;
+
+            result = ConvertToPrimaryExchangeRateCurrency(result, sourceCurrencyCode);
+            result = ConvertFromPrimaryExchangeRateCurrency(result, targetCurrencyCode);
             return result;
         }
 
@@ -247,20 +270,21 @@ namespace Nop.Services.Directory
         public virtual decimal ConvertToPrimaryExchangeRateCurrency(decimal amount, Currency sourceCurrencyCode)
         {
             if (sourceCurrencyCode == null)
-                throw new ArgumentNullException("sourceCurrencyCode");
+                throw new ArgumentNullException(nameof(sourceCurrencyCode));
 
             var primaryExchangeRateCurrency = GetCurrencyById(_currencySettings.PrimaryExchangeRateCurrencyId);
             if (primaryExchangeRateCurrency == null)
                 throw new Exception("Primary exchange rate currency cannot be loaded");
 
-            decimal result = amount; 
-            if (result != decimal.Zero && sourceCurrencyCode.Id != primaryExchangeRateCurrency.Id)
-            {
-                decimal exchangeRate = sourceCurrencyCode.Rate;
-                if (exchangeRate == decimal.Zero)
-                    throw new NopException(string.Format("Exchange rate not found for currency [{0}]", sourceCurrencyCode.Name));
-                result = result / exchangeRate;
-            }
+            var result = amount;
+            if (result == decimal.Zero || sourceCurrencyCode.Id == primaryExchangeRateCurrency.Id)
+                return result;
+
+            var exchangeRate = sourceCurrencyCode.Rate;
+            if (exchangeRate == decimal.Zero)
+                throw new NopException($"Exchange rate not found for currency [{sourceCurrencyCode.Name}]");
+            result = result / exchangeRate;
+
             return result;
         }
 
@@ -273,20 +297,21 @@ namespace Nop.Services.Directory
         public virtual decimal ConvertFromPrimaryExchangeRateCurrency(decimal amount, Currency targetCurrencyCode)
         {
             if (targetCurrencyCode == null)
-                throw new ArgumentNullException("targetCurrencyCode");
+                throw new ArgumentNullException(nameof(targetCurrencyCode));
 
             var primaryExchangeRateCurrency = GetCurrencyById(_currencySettings.PrimaryExchangeRateCurrencyId);
             if (primaryExchangeRateCurrency == null)
                 throw new Exception("Primary exchange rate currency cannot be loaded");
 
-            decimal result = amount;
-            if (result != decimal.Zero && targetCurrencyCode.Id != primaryExchangeRateCurrency.Id)
-            {
-                decimal exchangeRate = targetCurrencyCode.Rate;
-                if (exchangeRate == decimal.Zero)
-                    throw new NopException(string.Format("Exchange rate not found for currency [{0}]", targetCurrencyCode.Name));
-                result = result * exchangeRate;
-            }
+            var result = amount;
+            if (result == decimal.Zero || targetCurrencyCode.Id == primaryExchangeRateCurrency.Id)
+                return result;
+
+            var exchangeRate = targetCurrencyCode.Rate;
+            if (exchangeRate == decimal.Zero)
+                throw new NopException($"Exchange rate not found for currency [{targetCurrencyCode.Name}]");
+            result = result * exchangeRate;
+
             return result;
         }
 
@@ -299,20 +324,10 @@ namespace Nop.Services.Directory
         public virtual decimal ConvertToPrimaryStoreCurrency(decimal amount, Currency sourceCurrencyCode)
         {
             if (sourceCurrencyCode == null)
-                throw new ArgumentNullException("sourceCurrencyCode");
+                throw new ArgumentNullException(nameof(sourceCurrencyCode));
 
             var primaryStoreCurrency = GetCurrencyById(_currencySettings.PrimaryStoreCurrencyId);
-            if (primaryStoreCurrency == null)
-                throw new Exception("Primary store currency cannot be loaded");
-
-            decimal result = amount;
-            if (result != decimal.Zero && sourceCurrencyCode.Id != primaryStoreCurrency.Id)
-            {
-                decimal exchangeRate = sourceCurrencyCode.Rate;
-                if (exchangeRate == decimal.Zero)
-                    throw new NopException(string.Format("Exchange rate not found for currency [{0}]", sourceCurrencyCode.Name));
-                result = result / exchangeRate;
-            }
+            var result = ConvertCurrency(amount, sourceCurrencyCode, primaryStoreCurrency);
             return result;
         }
 
@@ -328,45 +343,9 @@ namespace Nop.Services.Directory
             var result = ConvertCurrency(amount, primaryStoreCurrency, targetCurrencyCode);
             return result;
         }
-       
 
-        /// <summary>
-        /// Load active exchange rate provider
-        /// </summary>
-        /// <returns>Active exchange rate provider</returns>
-        public virtual IExchangeRateProvider LoadActiveExchangeRateProvider()
-        {
-            var exchangeRateProvider = LoadExchangeRateProviderBySystemName(_currencySettings.ActiveExchangeRateProviderSystemName);
-            if (exchangeRateProvider == null)
-                exchangeRateProvider = LoadAllExchangeRateProviders().FirstOrDefault();
-            return exchangeRateProvider;
-        }
+        #endregion
 
-        /// <summary>
-        /// Load exchange rate provider by system name
-        /// </summary>
-        /// <param name="systemName">System name</param>
-        /// <returns>Found exchange rate provider</returns>
-        public virtual IExchangeRateProvider LoadExchangeRateProviderBySystemName(string systemName)
-        {
-            var descriptor = _pluginFinder.GetPluginDescriptorBySystemName<IExchangeRateProvider>(systemName);
-            if (descriptor != null)
-                return descriptor.Instance<IExchangeRateProvider>();
-
-            return null;
-        }
-
-        /// <summary>
-        /// Load all exchange rate providers
-        /// </summary>
-        /// <returns>Exchange rate providers</returns>
-        public virtual IList<IExchangeRateProvider> LoadAllExchangeRateProviders()
-        {
-            var exchangeRateProviders = _pluginFinder.GetPlugins<IExchangeRateProvider>();
-            return exchangeRateProviders
-                .OrderBy(tp => tp.PluginDescriptor)
-                .ToList();
-        }
         #endregion
     }
 }
